backend/requirements.txt
google-adk
litellm
fastapi
uvicorn
python-dotenv
httpx
tavily-python>=0.3.0
mcp>=1.0.0
pyinstaller>=6.0.0
